% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bes.R
\name{turnout_common}
\alias{turnout_common}
\title{Calculate estimated election turnout by demographic characteristic}
\usage{
turnout_common(dgraphic = c("age", "gender", "ethnicity", "religion",
  "education"), sample = FALSE, percent = TRUE, write = FALSE)
}
\arguments{
\item{dgraphic}{The demographic characteristic of interest. Must be
defined as one of the following values: age, gender, ethnicity,
religion, education. Must be defined between quotation marks.}

\item{sample}{A boolean which if TRUE returns the unweighted sample size.
FALSE returns weighted sample size. Default value is FALSE.
\code{percent} must be FALSE to show sample sizes.}

\item{percent}{A boolean which if FALSE does not calculate turnout
percentages. Default value is TRUE.}

\item{write}{A boolean which if TRUE creates a CSV in working directory.
Default value is FALSE.}
}
\value{
A tibble or CSV containing turnout estimates for selected
  demographic.
}
\description{
Calculates estimated election turnout by commonly asked demographic
characteristic for validated British Election Study 2017 face to face
survey data only.
}
