% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bes.R
\name{turnout_query}
\alias{turnout_query}
\title{Calculate estimated election turnout by demographic characteristic}
\usage{
turnout_query(data, dgraphic, vote, wt, validated = TRUE, result = 0,
  sample = FALSE, percent = TRUE, write = FALSE)
}
\arguments{
\item{data}{A dataframe containing
\href{http://www.britishelectionstudy.com/}{British Election Study}
internet or face to face survey data.}

\item{dgraphic}{The name of the variable for the demographic characteristic
of interest. Must be defined between quotation marks.}

\item{vote}{The name of the variable which captures response from suvey
respondent to the question if they voted or not in the election of
interest. Must be defined between quotation marks.}

\item{wt}{The name of the weight variable to be used in analysis. Must
be defined between quotation marks.}

\item{validated}{Use when survey respondent's vote registeration has not
been validated by British Election Study. A boolean which if FALSE
filters \code{data} to show those survey respondents who are over the age
of 18 and say they are registered to vote. Removes 'Dont Know' responses.
If FALSE \code{result} must be defined. Default value is TRUE.}

\item{result}{Use when \code{validated} is FALSE. A numeric between 0 and 1
for the known turnout percentage for the election in Great Britain.
Default value is 0.}

\item{sample}{A boolean which if TRUE returns the unweighted sample size.
FALSE returns weighted sample size. Default value is FALSE.
If \code {validated} is FALSE then \code{result} must be default.
\code{percent} must be FALSE to show sample sizes.}

\item{percent}{A boolean which if FALSE does not calculate turnout
percentages. Default value is TRUE.}

\item{write}{A boolean which if TRUE creates a CSV in working directory.
Default value is FALSE.}
}
\value{
A tibble or CSV containing turnout estimates for selected
  demographic.
}
\description{
Calculates estimated election turnout by demographic characteristic for
British Election Study survey data.
}
